Rename RunInBackground method and move HealthcheckResponse to api.go
